backend/pyproject.toml
[project]
name = "dlcrc-backend"
version = "1.0.0"
description = "Decision-Linked Competitive Research Copilot - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.39.0",
    "httpx>=0.25.2",
    "beautifulsoup4>=4.12.2",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "greenlet>=3.0.0",
    "pypdf>=4.0.0",
]

[dependency-groups]
dev = []
